split cdk resources
