+++ .github/workflows/deploy-pr-environment.yml
@@ -55,7 +55,7 @@ jobs:
       - name: Get CloudFront URL
         id: get-url
         run: |-
-          URL=$(cdk --app cdk.out --context envName=pr${{ github.event.number }} --outputs-file outputs.json deploy --require-approval never 2>/dev/null && cat outputs.json | jq -r '.[] | select(.CloudFrontUrl) | .CloudFrontUrl' || echo "Not available yet")
+          URL=$(cdk --app cdk.out --outputs-file outputs.json deploy --require-approval never 2>/dev/null && cat outputs.json | jq -r '.[] | select(.CloudFrontUrl) | .CloudFrontUrl' || echo "Not available yet")
           echo "CLOUDFRONT_URL=$URL" >> $GITHUB_OUTPUT
       - name: Find existing comment
         id: find-comment
